requirements.txt
anyio==4.0.0
asgiref==3.7.2
click==8.1.7
colorama==0.4.6
Django==4.2.7
exceptiongroup==1.1.3
graphql-core==3.2.3
h11==0.14.0
idna==3.4
libcst==1.1.0
markdown-it-py==3.0.0
mdurl==0.1.2
mypy-extensions==1.0.0
Pygments==2.17.1
python-dateutil==2.8.2
python-multipart==0.0.6
PyYAML==6.0.1
rich==13.7.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.32.0.post1
strawberry-graphql==0.215.0
strawberry-graphql-django==0.25.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
uvicorn==0.24.0.post1
